Modularize the exploratory data analysis portion of the code
Fixed Error Messages:
...
  File "/.../nf_streamlit/app/views/login.py", line 106, in <module>
    from data import eda
ModuleNotFoundError: No module named 'data'

...
FileNotFoundError: [Errno 2] No such file or directory: '../syn5522627/SYNAPSE_METADATA_MANIFEST.tsv'
  File "/.../nf_streamlit/app/views/data.py", line 124, in eda
    df_ratios = calculate_fit_ratios(df_compounds, dfs_drc, den_sis, num_sis)
UnboundLocalError: local variable 'den_sis' referenced before assignment

...
  File "/.../nf_streamlit/app/views/data.py", line 11, in <module>
    from ..syn5522627 import R_COLS, C_COLS, FILE_HIDE_COLS
ImportError: attempted relative import beyond top-level package

...
    login.loginapp()
AttributeError: partially initialized module 'views.login' has no attribute 'loginapp' (most likely due to a circular import)

...
    from views.login import loginapp
ImportError: cannot import name 'loginapp' from partially initialized module 'views.login' (most likely due to a circular import) (/.../nf_streamlit/app/views/login.py)

...
  File "/.../nf_streamlit/app/views/data.py", line 15, in <module>
    from .syn5522627 import R_COLS, C_COLS, FILE_HIDE_COLS
ModuleNotFoundError: No module named 'views.syn5522627'

References:
Introduction to the Python Pathlib Module
https://stackabuse.com/introduction-to-the-python-pathlib-module/

Python - Import from parent directory - GeeksforGeeks
https://www.geeksforgeeks.org/python-import-from-parent-directory/

How do I use relative paths in Python? • GITNUX
https://blog.gitnux.com/code/python-relative-path/

ImportError: attempted relative import with no known parent package
https://iq-inc.com/importerror-attempted-relative-import/

python - How to fix "Attempted relative import in non-package" even with __init__.py - Stack Overflow
https://stackoverflow.com/questions/11536764/how-to-fix-attempted-relative-import-in-non-package-even-with-init-py/
https://stackoverflow.com/a/27876800
https://stackoverflow.com/questions/11536764/how-to-fix-attempted-relative-import-in-non-package-even-with-init-py/27876800#27876800
